+++ features/support/app_pages/iva/military_email.rb
@@ -7,19 +7,31 @@ class MilitaryEmail < IDmeBase
   include ErrorMessages
 
 
-  def verify(affiliation:, populate: true, type:)
+  def verify(affiliation:, populate: true, type: nil)
     find("[data-option=#{container_attribute}]").find(".verification-header").click
     populate_affiliation(affiliation)
 
     if populate
 
-      build_fake_info
+      data = data_for(:mil_email) #info for duplicate user
+      build_fake_info #info for unique and denied users
 
       case type
       when "unique"
-        populate_fields(email: @unique_email)
+        populate_fields(email: @unique_email,
+                        fname: @fake_first_name,
+                        lname: @fake_last_name,
+                        dob:   @dob)
       when "denied"
-        populate_fields(email: "farooq@id.me")
+        populate_fields(email: data.fetch("denied_email"),
+                        fname: @fake_first_name,
+                        lname: @fake_last_name,
+                        dob:   @dob)
+      when "duplicate"
+        populate_fields(email: data.fetch("dupe_email"),
+                        fname: data.fetch("dupe_fname"),
+                        lname: data.fetch("dupe_lname"),
+                          dob: data.fetch("dupe_dob"))
       end
 
       if ["Service Member", "Military Supporter"].include?(affiliation)
@@ -30,7 +42,7 @@ class MilitaryEmail < IDmeBase
         %w(first_name last_name).each do |field|
           fill_in field, :with => Faker::Name.send(field)
         end
-        2.times {fill_in "birth_date", :with => "01/01/1985"}
+        2.times {fill_in "birth_date", :with => @dob}
         select_option(container_attribute, "#s2id_service_subgroup_id", "Veteran")
       end
 
@@ -46,15 +58,16 @@ class MilitaryEmail < IDmeBase
   def build_fake_info
     @fake_first_name = Faker::Name.first_name
     @fake_last_name = Faker::Name.last_name
+    @dob = Faker::Date.birthday.strftime("%m%d%Y")
     @unique_email = @fake_last_name+"#{rand(6 ** 8)}"+"@id.me"
   end
 
-  def populate_fields(email:)
-    fill_in "service_member_first_name", with: @fake_first_name
-    fill_in "service_member_last_name", with: @fake_last_name
-    2.times {fill_in "service_member_birth_date", with: "01/05/1985"}
+  def populate_fields(email:, fname:, lname:, dob:)
+    fill_in "service_member_first_name", with: fname
+    fill_in "service_member_last_name", with: lname
+    2.times {fill_in "service_member_birth_date", with: dob}
     %w(email email_confirmation).each do |field|
-      2.times {fill_in field, :with => email}
+      fill_in field, :with => email
     end
   end